% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_range.R
\name{det_range}
\alias{det_range}
\title{Determine the initial range for stability selection}
\usage{
det_range(data, model, times = 50, ...)
}
\arguments{
\item{data}{data frame}

\item{model}{lavaan output object.}

\item{times}{number of bootstrap samples used.}

\item{...}{Any additional arguments to pass to regsem() or cv_regsem().}
}
\value{
result the lambda values and the upper bound and lower bound of the interquartile range.
}
\description{
This function perform regsem on bootstrap samples to determine the initial range for stability selection. Interquartile range of the bootstrap optimal regularization amounts are uesd as the final range.
}
