% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_range_par.R
\name{det_range_par}
\alias{det_range_par}
\title{Determine the initial range for stability selection, parallel version}
\usage{
det_range_par(data, model, times = 50, ...)
}
\arguments{
\item{data}{data frame}

\item{model}{lavaan output object.}

\item{times}{number of bootstrap samples used.}

\item{...}{Any additional arguments to pass to regsem() or cv_regsem().}
}
\value{
result the lambda values and the upper bound and lower bound of the interquartile range.
}
\description{
This function perform regsem on bootstrap samples to determine the initial range for stability selection. Interquartile range of the bootstrap optimal regularization amounts are uesd as the final range. Parallelization is used to achieve faster performance.
}
