% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fiml_calc.R
\name{fiml_calc}
\alias{fiml_calc}
\title{Calculates the FIML objective function}
\usage{
fiml_calc(ImpCov, data, Areg, lambda, alpha, type, pen_vec, nvar)
}
\arguments{
\item{ImpCov}{expected covariance matrix.}

\item{data}{required dataset.}

\item{Areg}{A matrix with current parameter estimates.}

\item{lambda}{penalty value.}

\item{alpha}{mixture for elastic net.}

\item{type}{penalty type.}

\item{pen_vec}{vector of penalized parameters.}

\item{nvar}{number of variables.}
}
\description{
Calculates the FIML objective function
}
\examples{
\dontrun{
fiml_calc()
}
}
\keyword{fit}
\keyword{likelihood}
\keyword{maximum}
\keyword{regularization}

