% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_fun.R
\name{fit_fun}
\alias{fit_fun}
\title{Calculates the objective function values.}
\usage{
fit_fun(ImpCov, SampCov, Areg, lambda, alpha, type, pen_vec)
}
\arguments{
\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{Areg}{A matrix with current parameter estimates.}

\item{lambda}{penalty value.}

\item{alpha}{mixture for elastic net.}

\item{type}{penalty type.}

\item{pen_vec}{vector of penalized parameters.}
}
\description{
Calculates the objective function values.
}

