% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grad_fun.R
\name{grad_fun}
\alias{grad_fun}
\title{Calculates gradient using forward difference method.}
\usage{
grad_fun(par, ImpCov, SampCov, Areg, Sreg, A, S, F, A_fixed, A_est, S_fixed,
  S_est, lambda, type, pars_pen, I)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{Areg}{A matrix with current parameter estimates.}

\item{Sreg}{S matrix with current parameter estimates.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}

\item{A_fixed}{A matrix with fixed indicators.}

\item{A_est}{A matrix with parameter estimates.}

\item{S_fixed}{S matrix with fixed indicators.}

\item{S_est}{S matrix with parameter estimates.}

\item{lambda}{penalty value.}

\item{type}{penalty type.}

\item{pars_pen}{parameter indicators to be penalized.}

\item{I}{Identity matrix.}
}
\description{
Calculates gradient using forward difference method.
}

