% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grad_ram.R
\name{grad_ram}
\alias{grad_ram}
\title{Calculates the gradient vector based on von Oertzen & Brick, 2014}
\usage{
grad_ram(par, ImpCov, SampCov, Areg, Sreg, A, S, F, lambda, type, pars_pen,
  diff_par)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{Areg}{A matrix with current parameter estimates.}

\item{Sreg}{S matrix with current parameter estimates.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}

\item{lambda}{penalty value.}

\item{type}{penalty type.}

\item{pars_pen}{parameter indicators to be penalized.}

\item{diff_par}{parameter values to take deviations from.}
}
\description{
Calculates the gradient vector based on von Oertzen & Brick, 2014
}

