% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grad_ram_wMean.R
\name{grad_ram_wMean}
\alias{grad_ram_wMean}
\title{Calculates the gradient, accounting for means, Von Oertzen \& Brick, 2014}
\usage{
grad_ram_wMean(par, ImpCov, SampCov22, Areg, Sreg, A, S, F, SampMean, lambda,
  type, m, mu, m.pars)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{SampCov22}{Sample covariance matrix with means added.}

\item{Areg}{A matrix with current parameter estimates.}

\item{Sreg}{S matrix with current parameter estimates.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}

\item{SampMean}{sample mean vector.}

\item{lambda}{penalty value.}

\item{type}{penalty type.}

\item{m}{not sure.}

\item{mu}{not sure.}

\item{m.pars}{which parameters to penalize}
}
\description{
Calculates the gradient, accounting for means, Von Oertzen \& Brick, 2014
}

