% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hess_ram.R
\name{hess_ram}
\alias{hess_ram}
\title{Calculates the hessian matrix from von Oertzen & Brick, 2014}
\usage{
hess_ram(par, ImpCov, SampCov, Areg, Sreg, A, S, F)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{Areg}{A matrix with current parameter estimates.}

\item{Sreg}{S matrix with current parameter estimates.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}
}
\description{
Calculates the hessian matrix from von Oertzen & Brick, 2014
}

