% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hess_timoParallel.R
\name{hess_timoParallel}
\alias{hess_timoParallel}
\title{Parallel version to create Hessian based on Von Oertzen \& Brick, 2014}
\usage{
hess_timoParallel(par, ImpCov, SampCov, Areg, Sreg, A, S, F)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{Areg}{A matrix with current parameter estimates.}

\item{Sreg}{S matrix with current parameter estimates.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}
}
\description{
Parallel version to create Hessian based on Von Oertzen \& Brick, 2014
}
\examples{
\dontrun{
hess_timoParallel()
}
}
\keyword{derivative}
\keyword{gradient}

