% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hessian.R
\name{hessian}
\alias{hessian}
\title{Calculates the maximum likelihood fit function}
\usage{
hessian(par, ImpCov, SampCov, A, A_fixed, A_est, S, S_fixed, S_est, F, lambda,
  alpha, type)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{A}{A matrix with parameter labels.}

\item{A_fixed}{A matrix with fixed indicators.}

\item{A_est}{A matrix with parameter estimates.}

\item{S}{S matrix with parameter labels.}

\item{S_fixed}{S matrix with fixed indicators.}

\item{S_est}{S matrix with parameter estimates.}

\item{F}{F matrix.}

\item{lambda}{penalty value.}

\item{alpha}{value for elastic net penalty.}

\item{type}{penalty function type.}
}
\description{
Calculates the maximum likelihood fit function
}

