% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hessian_parallel.R
\name{hessian_parallel}
\alias{hessian_parallel}
\title{Parallel version of the finite forward difference method for the Hessian.}
\usage{
hessian_parallel(par, ImpCov, A, S, F, A_fixed, A_est, S_fixed, S_est, lambda,
  alpha, type)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}

\item{A_fixed}{A matrix with fixed indicators.}

\item{A_est}{A matrix with parameter estimates.}

\item{S_fixed}{S matrix with fixed indicators.}

\item{S_est}{S matrix with parameter estimates.}

\item{lambda}{penalty value.}

\item{alpha}{value for elastic net penalty.}

\item{type}{penalty function type.}
}
\description{
Parallel version of the finite forward difference method for the Hessian.
}

