% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optim_rj.R
\name{optim_rj}
\alias{optim_rj}
\title{An attempt at optimization written by ross jacobucci}
\usage{
optim_rj(start, func, grad, hess, tol, max.iter)
}
\arguments{
\item{start}{Starting values.}

\item{func}{function to compute objective value.}

\item{grad}{gradient vector.}

\item{hess}{hessian matrix.}

\item{tol}{absolute tolerance for convergence.}

\item{max.iter}{number of iterations to allow for convergence.}
}
\description{
An attempt at optimization written by ross jacobucci
}
\keyword{calc}
\keyword{optim}

