% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_mod.R
\name{pen_mod}
\alias{pen_mod}
\title{Penalized model syntax.}
\usage{
pen_mod(model, nm = NULL, pars_pen = NULL)
}
\arguments{
\item{model}{lavaan output object.}

\item{nm}{names(regsemOutput$coefficients).}

\item{pars_pen}{a vector of numbers corresponding to paths to be removed (same sequence as regsemOutput$coefficients).}
}
\value{
new.mod new model in lavaan syntax.
}
\description{
This function create a lavaan model syntax with paths corresponding to paremeters penalized to 0 removed.
}
