% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cvregsem}
\alias{plot.cvregsem}
\title{Plot function for cv_regsem}
\usage{
\method{plot}{cvregsem}(x, ..., pars = NULL, show.minimum = "BIC",
  col = NULL, type = "l", lwd = 3, lty = 1, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{x}{An x from cv_regsem.}

\item{...}{Other arguments.}

\item{pars}{Which parameters to plot}

\item{show.minimum}{What fit index to use}

\item{col}{A specification for the default plotting color.}

\item{type}{what type of plot should be drawn. Possible types are "p" for points, "l" for lines, or "b" for both}

\item{lwd}{line width}

\item{lty}{line type}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}
}
\description{
Plot function for cv_regsem
}
