% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.regsem}
\alias{plot.regsem}
\title{Plot function for cv_regsem}
\usage{
\method{plot}{regsem}(object, ...)
}
\arguments{
\item{object}{An object from regsem.}

\item{...}{Other arguments.}
}
\description{
Plot function for cv_regsem
}

