% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot function for cv_regsem}
\usage{
plot_cv(object, pars, color = TRUE, ...)
}
\arguments{
\item{object}{An object from cv_regsem.}

\item{pars}{Which parameters to plot}

\item{color}{Whether to plot in color}

\item{...}{Other arguments.}
}
\description{
Plot function for cv_regsem
}
