% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ram_calc.R
\name{ram_calc}
\alias{ram_calc}
\title{Calculates the fit function for von Oertzen & Brick, 2014}
\usage{
ram_calc(par, SampCov22, A, S, F, SampMean)
}
\arguments{
\item{par}{vector with parameters.}

\item{SampCov22}{Sample covariance matrix with means added.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}

\item{SampMean}{sample mean vector.}
}
\description{
Calculates the fit function for von Oertzen & Brick, 2014
}

