% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RAMmult.R
\name{RAMmult}
\alias{RAMmult}
\title{Take RAM matrices, multiplies, and returns Implied Covariance matrix.}
\usage{
RAMmult(par, A, S, F, A_fixed, A_est, S_fixed, S_est)
}
\arguments{
\item{par}{parameter estimates.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}

\item{A_fixed}{A matrix with fixed indicators.}

\item{A_est}{A matrix with parameter estimates.}

\item{S_fixed}{S matrix with fixed indicators.}

\item{S_est}{S matrix with parameter estimates.}
}
\description{
Take RAM matrices, multiplies, and returns Implied Covariance matrix.
}

