% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_grad_ram}
\alias{rcpp_grad_ram}
\title{Calculates the gradient vector based on Von Oertzen \& Brick, 2014}
\usage{
rcpp_grad_ram(par, ImpCov, SampCov, Areg, Sreg, A, S, F, lambda, type2,
  pen_vec, diff_par)
}
\arguments{
\item{par}{vector with parameters.}

\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{Areg}{A matrix with current parameter estimates.}

\item{Sreg}{S matrix with current parameter estimates.}

\item{A}{A matrix with parameter labels.}

\item{S}{S matrix with parameter labels.}

\item{F}{F matrix.}

\item{lambda}{penalty value.}

\item{type2}{penalty type.}

\item{pen_vec}{parameter indicators to be penalized.}

\item{diff_par}{parameter values to take deviations from.}
}
\description{
Calculates the gradient vector based on Von Oertzen \& Brick, 2014
}
