% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_grad_timo}
\alias{rcpp_grad_timo}
\title{Multiply a number by two}
\usage{
rcpp_grad_timo(par, ImpCov, SampCov, Areg, Sreg, A, S, F, I, lambda, type2,
  pen_vec, diff_par)
}
\description{
Multiply a number by two
}

