% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_quasi_calc}
\alias{rcpp_quasi_calc}
\title{Compute quasi Hessian}
\usage{
rcpp_quasi_calc(I, s, y, H)
}
\arguments{
\item{I}{identity matrix.}

\item{s}{s vector.}

\item{y}{y vector.}

\item{H}{previous Hessian.}
}
\description{
Compute quasi Hessian
}
