% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cvregsem.R
\name{summary.cvregsem}
\alias{summary.cvregsem}
\title{print information about cvregsem object}
\usage{
\method{summary}{cvregsem}(x, fit = "BIC", ...)
}
\arguments{
\item{x}{cv_regsem}

\item{fit}{Which fit index to use}
}
\description{
print information about cvregsem object
}
