% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.regsem}
\alias{summary.regsem}
\title{Summary results from regsem.}
\usage{
summary.regsem(object, ...)
}
\arguments{
\item{object}{An object from regsem.}

\item{...}{Other arguments.}
}
\description{
Summary results from regsem.
}

