% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace.R
\name{trace}
\alias{trace}
\title{Calculates the trace of a matrix.}
\usage{
trace(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
The sum of the diagonal entries in \code{mat}.
}
\description{
Calculates the trace of a matrix.
}
\examples{
\dontrun{
mm = matrix(c(1,2,3,4),2,2)
trace(mm)
}
}
\keyword{trace}

