% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_indices.R
\name{fit_indices}
\alias{fit_indices}
\title{Calculates the fit indices}
\usage{
fit_indices(model, CV = F, CovMat = NULL, data = NULL, n.boot = 100)
}
\arguments{
\item{model}{regsem model object.}

\item{CV}{cross-validation.}

\item{CovMat}{supply covariance matrix?}

\item{data}{supply the dataset?}

\item{n.boot}{number of bootstrap samples to take.}
}
\description{
Calculates the fit indices
}
\examples{
\dontrun{
fit_indices()
}
}
\keyword{chisq}
\keyword{fit}
\keyword{ncp}
\keyword{rmsea}

