% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi_optim.R
\name{multi_optim}
\alias{multi_optim}
\title{uses multiple starts to try and get convergence, running regsem.}
\usage{
multi_optim(lavObject, max.try = 10, lambda, lower = -Inf, upper = Inf,
  type, optMethod = "nlminb", gradFun = "timo", pars_pen = NULL,
  diff_par = NULL, hessFun = "timo")
}
\arguments{
\item{lavObject}{lavaan output object.}

\item{max.try}{number of starts to try before convergence.}

\item{lambda}{shrinkage parameter.}

\item{lower}{lower bound vector.}

\item{upper}{upper bound vector.}

\item{type}{type of penalty to use.}

\item{optMethod}{solver to use.}

\item{gradFun}{gradient function to use.}

\item{pars_pen}{parameter indicators to penalize.}

\item{diff_par}{parameter values to deviate from.}

\item{hessFun}{hessian function to use.}
}
\description{
uses multiple starts to try and get convergence, running regsem.
}
\examples{
\dontrun{
library(lavaan)
HS <- data.frame(scale(HolzingerSwineford1939[,7:15]))
mod <- '
f =~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8 + x9
'
outt = cfa(mod,HS)

fit1 <- multi_optim(outt,max.try=40,
                   lambda=0.1,type="lasso",
                   optMethod="nlminb", gradFun="timo")
}
}
\keyword{multiple}
\keyword{optim}

