% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cvregsem}
\alias{plot.cvregsem}
\title{Plot function for cv_regsem}
\usage{
\method{plot}{cvregsem}(
  x,
  ...,
  pars = NULL,
  show.minimum = "BIC",
  col = NULL,
  type = "l",
  lwd = 3,
  h_line = 0,
  lty = 1,
  xlab = NULL,
  ylab = NULL,
  legend.x = NULL,
  legend.y = NULL,
  legend.cex = 1,
  legend.bg = par("bg"),
  grey.out = FALSE
)
}
\arguments{
\item{x}{An x from cv_regsem.}

\item{...}{Other arguments.}

\item{pars}{Which parameters to plot}

\item{show.minimum}{What fit index to use}

\item{col}{A specification for the default plotting color.}

\item{type}{what type of plot should be drawn. Possible types are "p" for points, "l" for lines, or "b" for both}

\item{lwd}{line width}

\item{h_line}{Where to draw horizontal line}

\item{lty}{line type}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{legend.x}{x-coordinate of legend. See ?legend}

\item{legend.y}{y-coordinate of legend. See ?legend}

\item{legend.cex}{cex of legend. See ?legend}

\item{legend.bg}{legend background color. See ?legend}

\item{grey.out}{Add grey to background}
}
\description{
Plot function for cv_regsem
}
