% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_fit_fun}
\alias{rcpp_fit_fun}
\title{Calculates the objective function values.}
\usage{
rcpp_fit_fun(ImpCov, SampCov, type2, lambda, gamma, pen_vec, pen_diff,
  e_alpha)
}
\arguments{
\item{ImpCov}{expected covariance matrix.}

\item{SampCov}{Sample covariance matrix.}

\item{type2}{penalty type.}

\item{lambda}{penalty value.}

\item{gamma}{additional penalty for mcp and scad}

\item{pen_vec}{vector of penalized parameters.}

\item{pen_diff}{Vector of values to take deviation from.}

\item{e_alpha}{Alpha for elastic net}
}
\description{
Calculates the objective function values.
}
