% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cvregsem.R
\name{summary.cvregsem}
\alias{summary.cvregsem}
\title{print information about cvregsem object}
\usage{
\method{summary}{cvregsem}(object, ...)
}
\arguments{
\item{object}{cv_regsem object}

\item{...}{Additional arguments}
}
\description{
print information about cvregsem object
}
