% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regsem.R
\name{summary.regsem}
\alias{summary.regsem}
\title{Summary results from regsem.}
\usage{
\method{summary}{regsem}(object, ...)
}
\arguments{
\item{object}{An object from regsem.}

\item{...}{Other arguments.}
}
\value{
Details regarding convergence and fit
}
\description{
Summary results from regsem.
}
